:sparkles: Make random card order work!
