--- public/scripts/script.js
@@ -1,14 +1,0 @@
-import wensen from './wensen.js';
-
-// Fisher-Yates shuffle algorithm
-function shuffleArray(array) {
-    for (let i = array.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [array[i], array[j]] = [array[j], array[i]];
-    }
-    return array;
-}
-
-const shuffledWensen = shuffleArray(wensen);
-
-export { shuffledWensen };
