+++ router/routes.js
@@ -1,13 +1,23 @@
 import express from 'express';
 import dotenv from 'dotenv';
-import { shuffledWensen } from '../public/scripts/script.js';
+import wensen from './wensen.js';
 
 dotenv.config();
 
 const router = express.Router();
 
 router.get('/', (req, res) => {
-    console.log(shuffledWensen);
+
+    // Fisher-Yates shuffle algorithm
+    function shuffleArray(array) {
+        for (let i = array.length - 1; i > 0; i--) {
+            const j = Math.floor(Math.random() * (i + 1));
+            [array[i], array[j]] = [array[j], array[i]];
+        }
+        return array;
+    }
+
+    const shuffledWensen = shuffleArray(wensen);
     res.render('main', { layout: 'index', title: 'Home', wensen: shuffledWensen });
 });
