+++ views/main.hbs
@@ -27,7 +27,7 @@
                         <p>{{this.ambassadeurs}}</p>
                     </li>
                 </ul>
-                <p>Card ID: {{this.id}}</p>
+                {{!-- <p>Card ID: {{this.id}}</p> --}}
             </section>
 
         </a>
